Rename fieldset.py to fieldset_helper.py so we don't overwrite symbols.
