+++ uswds_forms/__init__.py
@@ -2,7 +2,7 @@ from .date import *  # NOQA
 from .errors import *  # NOQA
 from .form import *  # NOQA
 from .radio_and_checkbox import *  # NOQA
-from .fieldset import *  # NOQA
+from .fieldset_helper import *  # NOQA
 
 
 VERSION = '0.0.1'
